pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avivi-automation"
version = "0.1.0"
description = "Avivi Automation Manager — client launcher and master fleet server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "cryptography>=42.0.0",
    "httpx>=0.27",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "python-multipart>=0.0.9",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "PyQt6>=6.6",
    "pymysql>=1.1",
    "psycopg2-binary>=2.9",
    "psutil>=5.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["shared", "Avivi_Master", "Avivi_Client", "OpenClaw_Launcher"]

[project.scripts]
avivi-master = "avivi_master.__main__:main"
avivi-client = "avivi_client.__main__:main"
openclaw-launcher = "openclaw_launcher.__main__:main"
